Remove accessibleBreadcrumb mixin
