--- packages/web-app-files/src/mixins/accessibleBreadcrumb.ts
@@ -1,46 +1,0 @@
-import last from 'lodash-es/last'
-import { mapGetters } from 'vuex'
-
-export default {
-  computed: {
-    ...mapGetters('Files', ['totalFilesCount'])
-  },
-  methods: {
-    accessibleBreadcrumb_focusAndAnnounceBreadcrumb(sameRoute) {
-      const breadcrumbEl = document.getElementById('files-breadcrumb')
-      if (!breadcrumbEl) {
-        return
-      }
-      const activeBreadcrumb = last(breadcrumbEl.children[0].children)
-      const activeBreadcrumbItem = activeBreadcrumb.getElementsByTagName('button')[0]
-      if (!activeBreadcrumbItem) {
-        return
-      }
-
-      const itemCount = this.totalFilesCount.files + this.totalFilesCount.folders
-
-      const announcement = this.$gettextInterpolate(
-        this.$ngettext(
-          'This folder contains %{ amount } item.',
-          'This folder contains %{ amount } items.',
-          itemCount
-        ),
-        { amount: itemCount }
-      )
-
-      const translatedHint =
-        itemCount > 0 ? announcement : this.$gettext('This folder has no content.')
-
-      document.querySelectorAll('.oc-breadcrumb-sr').forEach((el) => el.remove())
-
-      const invisibleHint = document.createElement('p')
-      invisibleHint.className = 'oc-invisible-sr oc-breadcrumb-sr'
-      invisibleHint.innerHTML = translatedHint
-
-      activeBreadcrumb.append(invisibleHint)
-      if (sameRoute) {
-        activeBreadcrumbItem.focus()
-      }
-    }
-  }
-}
